+++ CHANGELOG.md
@@ -6,12 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Issue where first in-page nav item would duplicate the title of that chapter.
+
 ## [v3.0.0] - 2026-07-26
 
 ### Changed
 
 - Use custom classes for navigation data
-- Create interfaces for page navigation 
+- Create interfaces for page navigation
 - [BREAKING] default value for subItems in PageNavigation introduced as part of interface refactor
 - Create interface for parent title